Moved FPGA devices in modules dir
