+++ modules/fpga2014/include/fpga_encoder.hpp
@@ -50,4 +50,21 @@ public:
   s32 getValue(void);
 };
 
+//////////////////////////////////////////////////////////////////
+
+template<typename T, int ADDR>
+inline FpgaEncoder<T, ADDR>::FpgaEncoder(const char* name)
+  : Encoder(name) {
+}
+
+template<typename T, int ADDR>
+inline s32 FpgaEncoder<T, ADDR>::getValue(void) {
+  if(_inverse) {
+    return -(*(T*)ADDR);
+  }
+  else {
+    return *(T*)ADDR;
+  }
+}
+
 #endif//FPGA_ENCODER_HPP
