+++ modules/fpga2014/include/fpga_motor.hpp
@@ -16,11 +16,48 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef STM32_FPGA_ENCODER_HPP
-#define STM32_FPGA_ENCODER_HPP
+#ifndef FPGA_MOTOR_HPP
+#define FPGA_MOTOR_HPP
 
-#include "../../../common/device/encoder/fpga_encoder.hpp"
+#include <device/motor/motor.hpp>
+
+//! \brief A propulsion motor, used by Eirbot in 2014
+//! \param ADDR : the address where to set the motor pwm
+/*! 
+
+  This actuator is actually connected to a FPGA, which is viewed as an
+  external memory by the microcontroller. The address is set by the
+  user, so I hope you know how your FPGA works ;) !
+
+*/
+template<typename T, int ADDR>
+class FpgaMotor : public Motor {
+public:
+  FpgaMotor(const char* name);
+
+  void setValue(s32);
+};
+
+///////////////////////////////////////////////////////////////////
+
+#include <math/saturate.hpp>
+
+template<typename T, int ADDR>
+inline FpgaMotor<T, ADDR>::FpgaMotor(const char* name)
+  : Motor(name) {
+}
+
+template<typename T, int ADDR>
+inline void FpgaMotor<T, ADDR>::setValue(s32 val) {
+  const s32 cmd = Math::saturate(val, _min, _max);
+  if(_inverse) {
+    (*(T*)ADDR) = -cmd;
+  }
+  else {
+    (*(T*)ADDR) = cmd;
+  }
+}
 
 
 
-#endif//STM32_FPGA_ENCODER_HPP
+#endif//FPGA_MOTOR_HPP
